refactor(client): split auth.py into auth, oidc, and oidc_factory modules
- auth.py: protocols (TokenProvider, AsyncTokenProvider), StaticToken, AuthError
- oidc.py: OIDCTokenProvider, TokenSet, JWT helpers, OIDC discovery, TokenRefreshError
- oidc_factory.py: provider cache, config persistence callbacks, cross-process
  locking, resolve_oidc_provider

No re-exports — each consumer imports from the actual module.
