+++ packages/nemo_platform_plugin/src/nemo_platform_plugin/client/client.py
@@ -329,7 +329,7 @@ def _client_from_config(
     config_path: Path | str | None = None,
 ) -> _ClientT:
     """Shared implementation for NemoClient.from_config / AsyncNemoClient.from_config."""
-    from nemo_platform_plugin.client.auth import resolve_oidc_provider
+    from nemo_platform_plugin.client.oidc_factory import resolve_oidc_provider
     from nemo_platform_plugin.client.config.config import Config
     from nemo_platform_plugin.client.config.models import OAuthUser
